+++ runGame.bat
@@ -1,3 +1,3 @@
 @echo off
 
-start cmd.exe /c "python %~dp0\seeding_new.py"
+start cmd.exe /c "python %~dp0\seeding.py"
