+++ seeding.py
@@ -1,415 +1,595 @@
-import sys, a2s, yaml, time, subprocess, os
-from steam import game_servers as gs
+# required
+import sys, yaml, time, traceback, random
 from datetime import datetime as dt, timedelta
-from colors import colors
+from steam import game_servers as gs
+# project required
+import colors as c, hll_game, stopwatches as sw
 
-class Server:
-    def __init__(self):
-        self.desc = ""
-        self.server_ip = ""
-        self.connect_port = 0
-        self.query_port = 0
-        self.valid = False
+# Don't launch the game, join servers, or check their player lists
+# Otherwise, operates as if it did those things for testing
+debug_no_game = False
+# Print extra logs in a few places
+debug_extra_logs = False
 
-def process_exists(process_name):
-    progs = str(subprocess.check_output('tasklist'))
-    if process_name in progs:
-        return True
+
+def debug(log):
+    if debug_extra_logs:
+        print(f'{c.darkgrey}DEBUG : {log}{c.reset}')
+
+
+def try_parsing_time(possible_date, field):
+    for fmt in ('%I:%M %p', '%I:%M%p'):
+        try:
+            return dt.strptime(possible_date, fmt)
+        except ValueError:
+            pass
+    raise ValueError(f"Non-valid date format for field {field}: '{possible_date}'")
+
+
+def split_whitespace(string):
+    keywords = []
+
+    for split in string.split(' '):
+        split = split.strip()
+        if split and split not in keywords:
+            keywords.append(split)
+
+    return keywords
+
+
+print(f'{c.yellow}   ###############################   {c.reset}')
+print(f'{c.yellow}   ###   HLL Advanced Seeder   ###   {c.reset}')
+print(f'{c.yellow}   ###############################   {c.reset}')
+print()
+
+debug('Loading YAML')
+with open('seeding.yaml', 'r') as file:
+    seeding_yaml = yaml.safe_load(file)
+debug(f'{c.darkgrey}{seeding_yaml}{c.reset}\n')
+
+seeding_method = str(seeding_yaml["seeding_method"]).lower()
+seeding_endtime = try_parsing_time(seeding_yaml["seeding_endtime"], "seeding_endtime")
+seeding_minutes = int(seeding_yaml["seeding_minutes"])
+priority_monitor = bool(seeding_yaml["priority_monitor"])
+priority_monitor_endtime = try_parsing_time(seeding_yaml["priority_monitor_endtime"], "priority_monitor_endtime")
+servers = list(seeding_yaml["priority_servers"])
+seeded_player_limit = int(seeding_yaml["seeded_player_limit"])
+seeded_player_variability = int(seeding_yaml["seeded_player_variability"])
+server_query_rate = int(seeding_yaml["server_query_rate"])
+server_query_timeout = int(seeding_yaml["server_query_timeout"])
+query_timeout_limit = int(seeding_yaml["query_timeout_limit"])
+check_idle_kick = bool(seeding_yaml["check_idle_kick"])
+player_name = seeding_yaml["player_name"]
+perpetual_mode = bool(seeding_yaml["perpetual_mode"])
+perpetual_max_servers = int(seeding_yaml["perpetual_max_servers"])
+perpetual_min_players = int(seeding_yaml["perpetual_min_players"])
+ignore_name_contains = list(seeding_yaml["ignore_name_contains"])
+
+start_datetime = dt.today()
+stop_datetime = start_datetime
+if seeding_method == "endtime":
+    stop_datetime = start_datetime.replace(hour=seeding_endtime.hour, minute=seeding_endtime.minute, second=0,
+                                           microsecond=0)
+    if stop_datetime < start_datetime:
+        stop_datetime += timedelta(days=1)
+elif seeding_method == "minutes":
+    stop_datetime += timedelta(minutes=seeding_minutes)
+else:
+    print(f'{c.red}Invalid seeding_method. Check yaml and try again.{c.reset}')
+    input("Press enter to exit")
+    sys.exit()
+
+plan_runtime = (stop_datetime - start_datetime).total_seconds()
+plan_runtime_str = time.strftime("%Hh %Mm %Ss", time.gmtime(plan_runtime))
+
+stop_priority_datetime = start_datetime.replace(hour=priority_monitor_endtime.hour, minute=priority_monitor_endtime.minute, second=0,
+                                           microsecond=0)
+if stop_priority_datetime < start_datetime:
+    stop_priority_datetime += timedelta(days=1)
+
+plan_prioritytime = (stop_priority_datetime - start_datetime).total_seconds()
+plan_prioritytime_str = time.strftime("%Hh %Mm %Ss", time.gmtime(plan_prioritytime))
+
+print(f'{c.yellow}Summary{c.reset}')
+print(f'Run method        : {c.lightblue}{seeding_method}{c.reset} or when done')
+print(f'Start time        : {c.lightblue}{start_datetime}{c.reset}')
+print(f'Plan end time     : {c.lightblue}{stop_datetime}{c.reset} or {c.darkgrey}{plan_runtime_str} from start{c.reset}')
+print(f'Perpetual mode    : {c.green if perpetual_mode else c.darkgrey}{perpetual_mode}{c.reset}')
+print(f'Priority monitor  : {c.green if priority_monitor else c.darkgrey}{priority_monitor}{c.reset}')
+if priority_monitor:
+    print(f'Priority end time : {c.lightblue}{stop_priority_datetime}{c.reset} or {c.darkgrey}{plan_prioritytime_str} from start{c.reset}')
+print(f'Servers listed    : {c.lightblue}{len(servers)}{c.reset}')
+print()
+
+do_steam_search = perpetual_mode
+for server in servers:
+    if "steam_search" in server.keys():
+        do_steam_search = True
+        break
+
+steam_servers = {}
+if do_steam_search:
+    print(f'{c.yellow}Searching steam server list{c.reset}')
+    for server_addr in gs.query_master(r'\appid\686810', max_servers=1000):
+        value = f'{c.lightgrey}{len(steam_servers)} servers{c.reset}'
+        print("\r{0}".format(value), end='')
+        try:
+            info = gs.a2s_info(server_addr, timeout=2)
+
+            steam_servers[server_addr] = info
+        except Exception as err:
+            debug(f"\n{c.red}Unexpected A {err=}, {type(err)=}{c.reset}")
+    value = f'{c.lightgrey}{len(steam_servers)} servers{c.reset}'
+    print("\r{0}".format(value), end='\n')
+    print()
+
+# [ {"server_addr": (ip, port), "info": {a2s_info}}, ... ]
+priority_servers = []
+# [ (ip, port), ... ]
+server_queue = []
+# [ (ip, port), ... ]
+previously_joined = []
+
+
+def is_priority_server(server_addr):
+    for server in priority_servers:
+        if server_addr == server["server_addr"]:
+            return True
+    return False
+
+
+def get_priority_config(server_addr):
+    for server in priority_servers:
+        if server_addr == server["server_addr"]:
+            return server["config"]
+
+
+# server_addr   - (ip, port)
+# server_info   - {a2s_info}
+def should_server_queue(server_addr, server_info, min_players=0, name_ignore=None, verify_name=None,
+                        check_playercount=True):
+    if verify_name is None:
+        verify_name = []
+    if name_ignore is None:
+        name_ignore = []
+    queue = True
+    reasons = []
+    # Shouldn't happen but doesn't hurt to check
+    if server_info["game"] != "Hell Let Loose" or server_info["game_id"] != 686810:
+        queue = False
+        reasons.append("Server not HLL")
+    # Already queued
+    if server_addr in server_queue:
+        queue = False
+        reasons.append("Already queued")
+    # Previously joined
+    if server_addr in previously_joined:
+        queue = False
+        reasons.append("Previously joined")
+    # Only real 100 player servers. Not 64 player Bob the Builder
+    if server_info["max_players"] != 100:
+        queue = False
+        reasons.append("Not 100 max")
+    # Servers with no password
+    if server_info["visibility"] != 0:
+        queue = False
+        reasons.append("Passworded")
+    # Server name doesn't contain specified keywords
+    for word in name_ignore:
+        if str(word).lower() in str(server_info["name"]).lower():
+            queue = False
+            reasons.append(f"Name has [{word}]")
+    # Verify name does contain a specific word
+    for word in verify_name:
+        if str(word).lower() not in str(server_info["name"]).lower():
+            queue = False
+            reasons.append(f"Name missing [{word}]")
+    # Player count is within desired range
+    if check_playercount and (server_info["players"] < min_players or server_info["players"] >= seeded_player_limit):
+        queue = False
+        reasons.append(f"Not in range [{min_players},{seeded_player_limit}]")
+
+    return {"queue": queue, "reasons": reasons}
+
+
+print(f'{c.yellow}Checking/queueing priority servers{c.reset}')
+for try_server in servers:
+    if "steam_search" in try_server.keys():
+        potential_add = []
+        for server_addr in steam_servers:
+            info = steam_servers[server_addr]
+
+            check = should_server_queue(server_addr, info, verify_name=split_whitespace(try_server["steam_search"]),
+                                        check_playercount=False)
+
+            already_priority = is_priority_server(server_addr)
+
+            if check["queue"] and not already_priority:
+                potential_add.append((info["players"], server_addr, info))
+            else:
+                pass
+
+        potential_add.sort(key=lambda a: a[0], reverse=True)
+
+        if len(potential_add) > 0:
+            server_addr = potential_add[0][1]
+            info = potential_add[0][2]
+
+            print(
+                f'Priority server : {c.lightblue}{str(server_addr).ljust(27)}{c.reset}{c.darkgrey}{info["name"]}{c.reset}')
+            priority_servers.append({"server_addr": server_addr, "info": info, "config": try_server})
+
+    elif "address" in try_server.keys():
+        server_ip = str(try_server["address"].split(':')[0])
+        server_port = int(try_server["address"].split(':')[1])
+
+        query_port_offset = [15, 17890, 19238, 17670, 18238, 20127]
+        potential_query_port = [server_port]  # incase already using query port (battlemetrics copied)
+        for offset in query_port_offset:
+            potential_query_port.append(server_port + offset)
+
+        valid_port = False
+        for port in potential_query_port:
+            try:
+                server_addr = (server_ip, port)
+
+                info = gs.a2s_info(server_addr, timeout=1)
+                valid_port = True
+
+                check = should_server_queue(server_addr, info, verify_name=split_whitespace(try_server["verify"]),
+                                            check_playercount=False)
+
+                if server_addr not in steam_servers:
+                    steam_servers[server_addr] = info
+
+                if check["queue"]:
+                    print(
+                        f'Priority server : {c.lightblue}{str(server_addr).ljust(27)}{c.reset}{c.darkgrey}{info["name"]}{c.reset}')
+                    print(
+                        f'Priority seeding : {c.lightblue}{str(server_addr).ljust(27)}{c.reset}{c.darkgrey}{info["name"]}{c.reset}')
+
+                else:
+                    pass
+
+                break
+            except:
+                continue
+        if valid_port:
+            # print(f'Could not find valid query port')
+            pass
     else:
-        return False
+        continue
+print()
+
+# If not monitoring priority servers, queue seeding ones at runtime and go into perpetual after
+if not priority_monitor:
+    for server in priority_servers:
+        server_addr = server["server_addr"]
+        info = server["info"]
+        config = server["config"]
+
+        min_players = 0
+        if "min_players" in config:
+            min_players = int(config["min_players"])
+
+        check = should_server_queue(server_addr, info, min_players=min_players, check_playercount=True)
+
+        status_str = f'[{c.green}{str(info["players"])}{c.reset}/{c.green}{str(info["max_players"])}{c.reset}]'.rjust(
+            27)
+
+        if check["queue"]:
+            print(
+                f'{c.green}Seeding{c.reset} {status_str} : {c.lightblue}{str(server_addr).ljust(27)}{c.reset}{c.darkgrey}{info["name"]}{c.reset}')
+            server_queue.append(server_addr)
+        elif info["players"] >= seeded_player_limit:
+            print(
+                f'{c.darkgrey}Seeded{c.reset} {status_str} : {c.lightgrey}{str(server_addr).ljust(27)}{c.reset}{c.darkgrey}{info["name"]}{c.reset}')
+        else:
+            print(
+                f'{c.darkgrey}Skip{c.reset}   {status_str} : {c.lightgrey}{str(server_addr).ljust(27)}{c.reset}{c.darkgrey}{info["name"]}{c.reset}')
+            print(f'  {c.darkgrey}Skip reason(s){c.reset} : {c.darkgrey}{check["reasons"]}{c.reset}')
+    print()
+
+if not hll_game.is_running():
+    print(f'{c.lightgrey}Launching game...{c.reset}')
+
+    if not debug_no_game:
+        hll_game.launch_and_wait()
+else:
+    print(f'{c.lightgrey}Game already running{c.reset}')
+print()
+
+printed_progress = False
 
 
-def process_kill(ProcessName):
-    os.system('taskkill /IM "' + ProcessName + '" /F')
-
-
-def isNowInTimePeriod(startTime, endTime, nowTime):
-    if startTime < endTime:
-        return nowTime >= startTime and nowTime <= endTime
-    else:
-        # Over midnight:
-        return nowTime >= startTime or nowTime <= endTime
-
-
-def seed_progress(current, total, max_poll_count, seed_start, force_newline=False):
-    bar_length = 25
-    fraction = min(1.0, current / total)
-
-    arrow = int(fraction * bar_length - 1) * '-' + '>'
-    padding = int(bar_length - len(arrow)) * ' '
-
-    elapsed = (time.time() - seed_start)
-    elapsed_str = time.strftime("%Hh %Mm %Ss", time.gmtime(elapsed))
-
-    dead_percent = ''
-    threshold = int(max_poll_count / 2)
-    diff = max_poll_count - info.player_count
-    thresh_diff = max_poll_count - threshold
-    if seeding_yaml["move_on_if_server_dead"] and max_poll_count > 9:
-        dead_fraction = min(1.0, diff / thresh_diff)
-
-        dead_percent = f'Dying: {colors.fg.orange}{diff}{colors.reset}/{colors.fg.orange}{thresh_diff}  {int(dead_fraction * 100)}{colors.reset}%{colors.reset}'
-
-    progress_done = current >= total or (
-                seeding_yaml["move_on_if_server_dead"] and max_poll_count > 9 and diff >= thresh_diff) or force_newline
-    ending = '\n' if progress_done else '\r'
-
-    sys.stdout.write("\033[K");  # clear entire line
-    print(
-        f'Seed progress: [{colors.fg.green}{arrow}{colors.reset}{padding}]  Status: {colors.fg.green}{current}{colors.reset}/{colors.fg.green}{total}  {int(fraction * 100)}{colors.reset}%  Elapsed: {colors.fg.green}{elapsed_str}{colors.reset}  {dead_percent}',
-        end=ending)
+# Prefix for print() statements
+def nl():
+    global printed_progress
+    prefix = '\n' if printed_progress else ''
+    printed_progress = False
+    return prefix
 
 
 try:
-    print('Loading seeding.yaml')
-    with open('seeding.yaml', 'r') as file:
-        seeding_yaml = yaml.safe_load(file)
-    print(f'{colors.fg.darkgrey}{seeding_yaml}{colors.reset}')
+    print(f'{c.yellow}Starting seeding process{c.reset}')
     print()
 
-    query_timeout = seeding_yaml["server_query_timeout"]
+    search_for_next = False
+    next_server = True
+    current_server = None
+    latest_info = None
+    sw.start("seeding")
+    server_type = None
+    timeouts = 0
 
-    perpetual_mode = False
-    valid_servers = []
-    dead_servers = []
-    perpetual_ignore = seeding_yaml["perpetual_ignore_server_name_contains"]
+    player_threshold = seeded_player_limit
+    player_minimum = 0
 
-    today = dt.today()
-    stop_seeding_time = dt.strptime(seeding_yaml["stop_seeding_time"], "%I:%M %p")
-    stop_seeding_datetime = today.replace(hour=stop_seeding_time.hour, minute=stop_seeding_time.minute, second=0,
-                                          microsecond=0)
-    if stop_seeding_datetime < today:
-        stop_seeding_datetime += timedelta(days=1)
-
-    print(f'Targeted seeding stop time: {colors.fg.orange}{stop_seeding_datetime}{colors.reset}')
-    print()
+    players_max_count = 0
+    players_join_count = 0
+    players_done_count = 0
 
 
-    def perpetual_seeding_search():
-        print()
-        print(
-            f'{colors.fg.orange}Ran out of servers. Perpetual seeding enabled, searching for servers to seed...{colors.reset}')
+    def server_check():
+        global next_server, current_server
 
-        # Note: no password filter doesn't seem to work
-        servers_to_add = []
-        for server_addr in gs.query_master(r'\appid\686810\password\0\full\1\empty\1', max_servers=500):
+        priority_server = None
+        if priority_monitor is True and not is_priority_server(current_server):
+            # typically less than a second for ~10 servers
+            priority_server = priority_server_check()
+
+            if priority_server is not None and priority_server is not current_server:
+                print()
+                print(f'{c.lightcyan}Priority server now seeding{c.reset}')
+                server_queue.insert(0, priority_server)
+                next_server = True
+
+        if current_server is None and len(server_queue) == 0 and perpetual_mode and priority_server is None:
+            max = perpetual_max_servers
+            if priority_monitor:
+                max = 1
+
+            # typically ~30 seconds for ~250 servers assuming early disqualify
+            sw.start("perpetual")
+            perpetual_servers = perpetual_search(max_servers=max)
+            debug(f'{nl()}{c.darkgrey}{sw.seconds("perpetual")}s perpetual search - {perpetual_servers}{c.reset}')
+
+            if perpetual_servers is not None and len(perpetual_servers) > 0:
+                for server in perpetual_servers:
+                    server_queue.append(server)
+                next_server = True
+            print()
+
+
+    # Keep checking priority server states and switch to them when needed
+    def priority_server_check():
+        for server in priority_servers:
+            server_addr = server["server_addr"]
+            config = server["config"]
+
             try:
-                info = a2s.info(server_addr, timeout=2)
+                global latest_info
+                latest_info = gs.a2s_info(server_addr, timeout=server_query_timeout)
+                server["info"] = latest_info
+                steam_servers[server_addr] = latest_info
 
-                if server_addr in dead_servers:
-                    continue
+                min_players = 0
+                if "min_players" in config:
+                    min_players = int(config["min_players"])
 
-                bad_name = False
-                for bad_text in perpetual_ignore:
-                    if str(bad_text).lower() in info.server_name:
-                        bad_name = True
+                check = should_server_queue(server_addr, latest_info, min_players=min_players, check_playercount=True)
 
-                if not info.password_protected and info.player_count < seeding_yaml[
-                    "seeded_player_limit"] and info.player_count >= seeding_yaml[
-                    "perpetual_min_players"] and info.max_players == 100 and not bad_name:
-                    servers_to_add.append((info.player_count, info, server_addr))
+                global current_server
+                if check["queue"] and server_addr not in server_queue and server_addr != current_server:
+                    return server_addr
             except:
                 continue
+        pass
 
-        # join most populated servers first
-        servers_to_add.sort(key=lambda a: a[0], reverse=True)
 
+    # Perpetual mode query server list again for seeding servers and queue up
+    def perpetual_search(max_servers=perpetual_max_servers):
+        print()
+        print(f'{c.lightcyan}Perpetual mode searching for more seeding servers...{c.reset}')
+
+        early_ignored = 0
+        potential_add = []
+        for server_addr in steam_servers:
+            early_check = should_server_queue(server_addr, steam_servers[server_addr],
+                                              name_ignore=ignore_name_contains,
+                                              check_playercount=False)
+            if not early_check["queue"]:
+                early_ignored += 1
+                # debug(f'{c.darkgrey}Early disqualify {server_addr} {early_check["reasons"]} {steam_servers[server_addr]["name"]}{c.reset}')
+                continue
+
+            try:
+                info = gs.a2s_info(server_addr, timeout=1)
+                steam_servers[server_addr] = info
+                check = should_server_queue(server_addr, info, min_players=perpetual_min_players,
+                                            name_ignore=ignore_name_contains)
+                if check["queue"]:
+                    potential_add.append((info["players"], server_addr))
+            except:
+                continue
+        potential_add.sort(key=lambda a: a[0], reverse=True)
+
+        debug(f'{c.darkgrey}{early_ignored} servers early ignored{c.reset}')
+
+        to_return = []
         i = 1
-        for server in servers_to_add:
-            if i > seeding_yaml["perpetual_max_servers"]:
+        for server in potential_add:
+            if i > max_servers:
                 break
-            info = server[1]
-            server_addr = server[2]
-
-            desc = info.server_name[0:10]
-
+            server_addr = server[1]
+            info = steam_servers[server_addr]
+            status_str = f'[{c.green}{str(info["players"])}{c.reset}/{c.green}{str(info["max_players"])}{c.reset}]'.rjust(
+                27)
             print(
-                f'{colors.fg.green}SEEDING (Queued) [ {desc} ] query_port={server_addr[1]}, status={info.player_count}/{info.max_players}{colors.reset}')
-            print(f'    {colors.fg.darkgrey}{info.server_name}{colors.reset}')
-
-            server = Server()
-            server.desc = desc
-            server.server_ip = server_addr[0]
-            server.query_port = server_addr[1]
-            server.valid = True
-
-            valid_servers.append(server)
+                f'{c.green}Queued{c.reset} {status_str} : {c.lightblue}{str(server_addr).ljust(27)}{c.reset}{c.darkgrey}{info["name"]}{c.reset}')
+            to_return.append(server_addr)
             i += 1
+        return to_return
 
 
-    # Check the servers listed to monitor
-    # Are they still valid? What's the query port?
-    print('Initial server check')
+    def seed_progress(current):
+        bar_length = 15
+        fraction = min(1.0, current / player_threshold)
+        arrow = int(fraction * bar_length - 1) * '-' + '>'
+        padding = int(bar_length - len(arrow)) * ' '
 
-    for try_server in seeding_yaml["seeding_list"]:
-        game_connect = try_server["connect"]
+        threshold = int(players_max_count / 2)
+        diff = players_max_count - current
+        thresh_diff = players_max_count - threshold
+        dead_fraction = min(1.0, diff / thresh_diff)
 
-        server_ip = game_connect.split(':')[0]
-        server_connect_port = int(game_connect.split(':')[1])
+        progress_bar = f'[{c.green}{arrow}{c.reset}{padding}]'
+        status_str = (f'Status: {c.darkgrey}{player_minimum}{c.reset}/{c.green}{current}{c.reset}/{c.green}{player_threshold}{c.reset}'
+                      f'  {c.green}{int(fraction * 100)}{c.reset}%')
+        elapsed_str = f'Elapsed: {c.green}{time.strftime("%Hh %Mm %Ss", time.gmtime(sw.seconds("seeding")))}{c.reset}'
+        dying_str = f'Dying: {c.orange}{diff}{c.reset}/{c.orange}{thresh_diff}  {int(dead_fraction * 100)}{c.reset}%{c.reset}'
+        timeout_str = "" if timeouts == 0 else f'Timeout: {c.red}{timeouts}{c.reset}/{c.red}{query_timeout_limit}{c.reset}'
 
-        query_port_offset = [
-            15,     # G-Portal
-            17890,  # GTX
-            19238,  # Low.MS
-            17670,  # ???       offset from fll.fi / WTH
-            18238,  # ???       offset from HLL Brasil
-            20127   # ???       offset from chinese servers
-        ]
-        potential_query_ports = []
-        for offset in query_port_offset:
-            potential_query_ports.append(server_connect_port + offset)
-        potential_query_ports.append(server_connect_port)  # incase accidentally using query in yaml
+        value = f'Seed progress {progress_bar}  {status_str}  {elapsed_str}  {dying_str}  {timeout_str}  '
+        print("\r{0}".format(value), end='')
+        global printed_progress
+        printed_progress = True
 
-        valid_server = False
-        quit_early = False
 
-        for query_port in potential_query_ports:
-            try:
-                info = a2s.info((server_ip, query_port), timeout=1)
-                desc = try_server["description"]
-
-                if seeding_yaml["verify_name"] and "verify_name" in try_server.keys() and try_server[
-                    "verify_name"] not in info.server_name:
-                    print(
-                        f'{colors.fg.red}INVALID [ {desc} ]. Server name did not contain keyword [ {try_server["verify_name"]} ]{colors.reset}')
-                    print(f'    {colors.fg.darkgrey}{info.server_name}{colors.reset}')
-                    quit_early = True
-                    break
-                if info.password_protected:
-                    print(f'{colors.fg.red}INVALID [ {desc} ] was password protected{colors.reset}')
-                    print(f'    {colors.fg.darkgrey}{info.server_name}{colors.reset}')
-                    quit_early = True
-                    break
-
-                valid_server = True
-
-                if info.player_count >= seeding_yaml["seeded_player_limit"]:
-                    print(
-                        f'{colors.fg.lightgrey}SEEDED (Ignoring) [ {desc} ] query_port={query_port}, status={info.player_count}/{info.max_players}{colors.reset}')
-                    print(f'    {colors.fg.darkgrey}{info.server_name}{colors.reset}')
-                    break
-                else:
-                    print(
-                        f'{colors.fg.green}SEEDING (Queued) [ {desc} ] query_port={query_port}, status={info.player_count}/{info.max_players}{colors.reset}')
-                    print(f'    {colors.fg.darkgrey}{info.server_name}{colors.reset}')
-
-                    server = Server()
-                    server.desc = try_server["description"]
-                    server.server_ip = server_ip
-                    server.connect_port = server_connect_port
-                    server.query_port = query_port
-                    server.valid = True
-
-                    valid_servers.append(server)
-
-                break
-            except:
-                continue
-
-        if not valid_server and not quit_early:
-            print(f'{colors.fg.red}INVALID [ {try_server["description"]} ] Could not find query port{colors.reset}')
-    print()
-    print()
-
-    if len(valid_servers) > 0 or seeding_yaml["perpetual_seed_from_steam"]:
-        if process_exists('HLL-Win64-Shipping.exe'):
-            print("Game already running, skipping launch and wait")
-            time.sleep(1)
+    while True:
+        time.sleep(1)
+        if next_server and len(server_queue) == 0 and not priority_monitor and not perpetual_mode:
             print()
+            print(f'Ran out of servers.')
+            # no servers and no ongoing processes, script done
+            break
+
+        # Attempt longer delays between queries
+        sleep = server_query_rate + (10 * max(0, timeouts))
+        time.sleep(sleep)
+
+        server_check()
+
+        if len(server_queue) >= 1 and next_server:
+            current_server = server_queue.pop(0)
+            info = steam_servers[current_server]
+            previously_joined.append(current_server)
+            next_server = False
+            players_max_count = 0
+            timeouts = 0
+            sw.start("seeding")
+            player_threshold = seeded_player_limit + random.randrange(0, seeded_player_variability)
+            player_threshold = min(player_threshold, info["max_players"])
+            server_type = "Priority" if is_priority_server(current_server) else "Perpetual"
             print()
-        else:
-            print("Launching game and waiting 60 seconds...")
-            subprocess.run("cmd /c start steam://run/686810")
-            time.sleep(60)
+            print(f'{nl()}{c.yellow}Monitoring Server ({server_type}){c.reset}')
+            printed_progress = False
+            print(f'{c.darkgrey}{info["name"]}{c.reset}')
+            print(f'Connecting {c.lightblue}{str(current_server).ljust(27)}{c.reset}')
+
+            player_minimum = 0
+            if is_priority_server(current_server):
+                config = get_priority_config(current_server)
+                if config is not None and "min_players" in config:
+                    player_minimum = int(config["min_players"])
+            else:
+                player_minimum = perpetual_min_players
+
+            if not debug_no_game:
+                sw.start("idle_check")
+                hll_game.join_server_addr(current_server)
+                time.sleep(15)
+
+        if dt.today() >= stop_datetime:
             print()
-            print()
+            print(f'{nl()}{c.yellow}Seeding stop time reached{c.reset}')
+            hll_game.kill()
+            hll_game.wait_until_dead()
+            break
+        elif priority_monitor and dt.today() >= stop_priority_datetime:
+            print(f'{nl()}{c.orange}Priority monitor stop time reached.{c.reset}')
+            priority_monitor = False
 
-        seed_start = time.time()
+            if is_priority_server(current_server):
+                current_server = None
+            continue
+        if current_server is None and len(server_queue) == 0:
+            debug(f'{nl()}{c.darkgrey}current_server is None and len(server_queue) == 0{c.reset}')
+            continue
+        elif current_server is None and len(server_queue) >= 1:
+            debug(f'{nl()}{c.darkgrey}current_server is None and len(server_queue) >= 1{c.reset}')
+            next_server = True
+            continue
 
-        seed_index = 0
-        monitor_start = False
-        monitor_start2 = False
-        tried_connect = False
-        exception_retry = 0
-        poll_count = 0  # times the script polled the current server
-        max_poll_count = 0  # max poll player count
-        players = None
-        next_reset = False
-        server = None
+        try:
+            latest_info = gs.a2s_info(current_server, timeout=server_query_timeout)
 
-        print("Starting seeding server rotation")
-        while True:
-            if next_reset:
-                seed_index += 1
-                exception_retry = 0
-                monitor_start = False
-                monitor_start2 = False
-                poll_count = 0
-                max_poll_count = 0
-                dead_servers.append((server.server_ip, server.query_port))
-                print(f'{colors.fg.darkgrey}{dead_servers}{colors.reset}')
-                players = None
+            players = latest_info["players"]
+            if players > players_max_count:
+                players_max_count = players
 
-                next_reset = False
+            seed_progress(players)
 
-            if seed_index >= len(valid_servers):
-                if seeding_yaml["perpetual_seed_from_steam"]:
-                    perpetual_seeding_search()
-                    perpetual_mode = True
+            if players >= player_threshold:
+                print(f'{nl()}{c.lightgreen}Seeded!{c.reset}')
+                current_server = None
+            elif is_priority_server(current_server) and players < player_minimum:
+                print(f'{nl()}{c.orange}Priority server below configured {player_minimum} players{c.reset}')
+                current_server = None
+            elif players < player_minimum and not is_priority_server(current_server):
+                print(f'{nl()}{c.orange}Perpetual server below {player_minimum} players{c.reset}')
+                current_server = None
+            elif players_max_count > player_minimum and players <= players_max_count / 2:
+                print(f'{nl()}{c.orange}Player count halved, server likely dying{c.reset}')
+                current_server = None
 
-                    if seed_index >= len(valid_servers):
-                        print(
-                            f'{colors.fg.orange}Failed to find more servers. Waiting a couple minutes...{colors.reset}')
-                        time.sleep(180)
-                else:
-                    break
+            if not debug_no_game:
+                if hll_game.did_game_crash():
+                    print(f'{nl()}{c.red}Game crashed{c.reset}')
+                    print(f'{nl()}{c.darkgrey}Relaunching game...{c.reset}')
+                    hll_game.relaunch_and_wait()
 
-            if dt.today() >= stop_seeding_datetime:
-                print()
-                print()
-                print(
-                    f'{colors.fg.yellow}Seeding stop time reached [ {stop_seeding_datetime} ]. Closing game.{colors.reset}')
+                    if current_server is not None:
+                        print(f'{nl()}Reconnecting {c.lightblue}{str(current_server).ljust(27)}{c.reset}')
+                        hll_game.join_server_addr(current_server)
+                        time.sleep(15)
 
-                process_kill('HLL-Win64-Shipping.exe')
-                break
-
-            server = valid_servers[seed_index]
-
-            if not monitor_start:
-                print()
-                print()
-                print(f'{colors.fg.yellow}Monitoring [ {server.desc} ]{colors.reset}')
-                tried_connect = False
-                monitor_start = True
-                seed_start = time.time()
-
-            try:
-                info = a2s.info((server.server_ip, server.query_port), timeout=query_timeout)
-
-                if seeding_yaml["move_on_if_kicked"]:
+                elif check_idle_kick and sw.seconds("idle_check") > 60:
                     try:
-                        players = gs.a2s_players((server.server_ip, server.query_port), timeout=query_timeout)
-                    except:
-                        players = None
+                        players = gs.a2s_players(current_server, timeout=server_query_timeout)
 
-                # Start checking player list one minute after connection attempt
-                if seeding_yaml["move_on_if_kicked"] and tried_connect and poll_count * seeding_yaml[
-                    "server_query_rate"] > 60 and players != None:
-                    name_present = False
-                    names = []
-                    for player in players:
-                        names.append(player["name"])
-                        if player["name"] == seeding_yaml["player_name"]:
-                            name_present = True
-                            break
+                        names = []
+                        for player in players:
+                            names.append(player["name"])
 
-                    if not name_present:
-                        seed_progress(info.player_count, seeding_yaml["seeded_player_limit"], max_poll_count,
-                                      seed_start, True)
+                        name_present = player_name.lower() in (string.lower() for string in names)
 
-                        names.sort()
-                        print(
-                            f'{colors.fg.red}{seeding_yaml["player_name"]} is no longer in the player list. Idle kicked?{colors.reset}')
-                        print(f'{colors.fg.darkgrey}{names}{colors.reset}')
+                        if not name_present:
+                            print(f'{nl()}{c.red}{player_name} is no longer in the player list. Idle kick?{c.reset}')
+                            print(f'{nl()}{c.darkgrey}Relaunching game...{c.reset}')
+                            hll_game.relaunch_and_wait()
+                            current_server = None
 
-                        # Unfortunately trying to join another server does not skip the message and only way to close it is to manually click the Ok button
-                        # So, killing and relaunching the game to join another server is the solution for now
-                        process_kill('HLL-Win64-Shipping.exe')
-                        time.sleep(5)
-                        print("Relaunching game and waiting 120 seconds...")
-                        time.sleep(60)
-                        subprocess.run("cmd /c start steam://run/686810")
-                        time.sleep(60)
-                        print()
-                        print()
+                    except Exception as err:
+                        pass
 
-                        next_reset = True
-                        time.sleep(30)
-                        continue
+        except Exception as err:
+            timeouts += 1  # not a timeout but still quit if it errors N times
 
-                if not monitor_start2:
-                    print(f'{colors.fg.darkgrey}{info.server_name}{colors.reset}')
-                    monitor_start2 = True
+            seed_progress(latest_info["players"])
 
-                if info.player_count > max_poll_count:
-                    max_poll_count = info.player_count
-                if seeding_yaml["move_on_if_server_dead"] and max_poll_count >= seeding_yaml[
-                    "perpetual_min_players"] and info.player_count <= max_poll_count / 2:
-                    seed_progress(info.player_count, seeding_yaml["seeded_player_limit"], max_poll_count, seed_start)
+            if timeouts >= query_timeout_limit:
+                print(f'{nl()}{c.red}Reached timeout limit{c.reset}')
+                current_server = None
 
-                    print(f'{colors.fg.orange}{server.desc} player count has halved, likely dead.{colors.reset}')
-                    print('Moving on.')
-                    next_reset = True
-                    continue
+            debug(f"{nl()}{c.red}Unexpected B {err=}, {type(err)=}{c.reset}")
+            continue
 
-                if seeding_yaml["perpetual_dead_if_empty"] and perpetual_mode and info.player_count < seeding_yaml[
-                    "perpetual_min_players"] - 1:
-                    seed_progress(info.player_count, seeding_yaml["seeded_player_limit"], max_poll_count, seed_start,
-                                  True)
-
-                    print(f'{colors.fg.orange}{server.desc} player count looks dead (perpetual mode).{colors.reset}')
-                    print('Moving on.')
-                    next_reset = True
-                    continue
-
-                if info.player_count >= seeding_yaml["seeded_player_limit"]:
-                    seed_progress(info.player_count, seeding_yaml["seeded_player_limit"], max_poll_count, seed_start)
-
-                    print(f'{server.desc} is seeded {info.player_count}/{info.max_players}')
-                    print('Moving on.')
-                    next_reset = True
-                    continue
-
-                if not tried_connect:
-                    connect = f"steam://connect/{server.server_ip}:{server.query_port}"
-                    command = f"cmd /c start {connect}"
-
-                    print('Connecting...', connect)
-                    subprocess.run(command)
-                    tried_connect = True
-                    time.sleep(5)
-
-                seed_progress(info.player_count, seeding_yaml["seeded_player_limit"], max_poll_count, seed_start)
-
-            except Exception as err:
-                print(f"\n{colors.fg.red}Unexpected {err=}, {type(err)=}{colors.reset}")
-                print(
-                    f"{colors.fg.red}Problem querying valid server {server.desc}. Retry {exception_retry + 1}/4{colors.reset}")
-
-                if exception_retry < 3:
-                    exception_retry += 1
-                    time.sleep(30)
-                else:
-                    print(f"{colors.fg.red}Failed to query 4 times. Moving on.")
-                    server.valid = False
-                    seed_index += 1
-                    exception_retry = 0
-                    monitor_start = False
-
-            if seed_index >= len(valid_servers):
-                if seeding_yaml["perpetual_seed_from_steam"]:
-                    perpetual_seeding_search()
-                    perpetual_mode = True
-
-                    if seed_index >= len(valid_servers):
-                        print(
-                            f'{colors.fg.orange}Failed to find more servers. Waiting a couple minutes...{colors.reset}')
-                        time.sleep(180)
-                else:
-                    break
-
-            poll_count += 1
-            time.sleep(seeding_yaml["server_query_rate"])
-
-        print()
-        print()
-
-        print("Seeding Done!")
-    else:
-        print("No servers to be seeded")
-
+    print()
+    print(f"{c.lightgreen}Seeding done!{c.reset}")
 except Exception as err:
-    print(f"Unexpected {err=}, {type(err)=}")
+    traceback.print_exc()
+    print(f"{nl()}{c.red}Unexpected A {err=}, {type(err)=}{c.reset}")
 
 print()
 input("Press enter to exit")
