--- seeding_new.yaml
@@ -1,78 +1,0 @@
-# Method ("endtime" or "minutes") to determine when to kill the game stop the script
-seeding_method: endtime
-seeding_endtime: 10:00 pm
-seeding_minutes: 720
-
-# Monitor priority servers when done or in perpetual mode.
-# Monitor helps if certain servers seed later than others
-priority_monitor: true
-priority_monitor_endtime: 10:00 pm
-# List of servers to seed
-# Configure servers to seed by "ip:port" or with keywords to find from the steam server list
-priority_servers: [
-  # always tries to seed early
-  {steam_search: saucymuffin beer},
-  {steam_search: glow's east},
-  {steam_search: glow's west},
-  {steam_search: outpost bhb},
-  {steam_search: soulsniper},
-  # only seed/join if others are on trying and leave if they leave
-  {
-    steam_search: brewdawg east,
-    min_players: 3
-  },
-  {
-    steam_search: chad stuart flare,
-    min_players: 4
-  },
-  {
-    steam_search: rotn rangers,
-    min_players: 6
-  },
-  {
-    steam_search: brewdawg west,
-    min_players: 3
-  },
-]
-# Example (ip:port) with optional name verification
-# Find ip:port from the Steam Server Browser or from BattleMetrics
-# Verify checks the server name contains all words listed split by whitespace
-#  {
-#    about: Saucymuffin,
-#    address: 192.169.95.146:9075,
-#    verify: BEER
-#  }
-#
-# Example (keywords search) split by whitespace, server must contain all words.
-# If multiple servers match, ordered by most populated not already queued and must fit seeding criteria
-# Watch for conflicts with {ignore_name_contains} property
-#  {steam_search: foo east}        for example will match: "FOO Server :: US East - blah blah blah"
-
-
-# Max player count before moving to the next server
-# 40 is technically seeded though it's best to stick around to at least 50-60
-seeded_player_limit: 50
-# If you are running multiple seed scripts/accounts, randomly add 0-N to the limit to mitigate scripts moving on at the same time.
-seeded_player_variability: 5
-# Delay between server poll in seconds
-server_query_rate: 15
-# Server poll timeout in seconds
-server_query_timeout: 15
-# How many timeouts before moving on when monitoring a server
-query_timeout_limit: 6
-# Check if name is in the player list
-check_idle_kick: false
-player_name: SodiumEnglish
-
-
-# Perpetually seed from the steam server list when done with priority servers
-# Help out the rest of the community!
-perpetual_mode: true
-# How many servers to queue for seeding (ordered from most to least populated)
-perpetual_max_servers: 2
-# Minimum players to consider seeding
-perpetual_min_players: 10
-# Do not seed servers with these terms in the name (case-insensitive)
-ignore_name_contains: [
-  cn, event, training, hll official
-]
